requirements.txt
requests=2.27.1
telebot=0.0.4
pymongo=4.3.3
bs4=0.0.1
python-binance=1.0.17